pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-fetch"
dynamic = ["version"]
description = "CLI tool for fetching AGENTS.md files from GitHub"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
authors = [
    { name = "FreeMarketaMilitia" },
]
maintainers = [
    { name = "FreeMarketaMilitia" },
]
keywords = [
    "cli", "github", "fetch", "agents", "documentation",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Utilities",
]
urls = { Repository = "https://github.com/example/agent-fetch" }
dependencies = [
    "typer>=0.9.0",
    "pyyaml>=6.0.0",
    "requests>=2.31.0",
    "questionary>=2.0.0",
    "rapidfuzz>=3.0.0",
    "rich>=13.0.0",
]

[project.scripts]
agentfetch = "agents_collector.cli.cli_main:app"

[tool.hatch.version]
path = "agents_collector/__init__.py"

[tool.hatch.build]
include = [
    "agents_collector",
    "README.md",
]

[tool.hatch.metadata]
allow-direct-references = true

# Optional: uv-specific configuration
[tool.uv]
dev-dependencies = []
